spectral: new package for spectral graph analysis
Initial commit to this is just a move of the Laplacian type from network.
