+++ graph/spectral/doc.go
@@ -1,0 +1,6 @@
+// Copyright ©2019 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package spectral provides graph spectral analysis functions.
+package spectral // import "gonum.org/v1/gonum/graph/spectral"
